Improved ripple effect
I have changed the ripple effect so that it is more in line with how it is presented by Google. The circle now expands to fill the button when the button has been clicked or pressed. I added a listener for a mouse release to detect when to kill the ripple.

http://i.imgur.com/KLKfbKX.gifv